refactor: Fix ci typecheck by removing donation overlay no-slug page file
